Move paths to a dedicated package
