+++ src/main/java/frc/robot/autonomous/paths/RightConeCubeHigh.java
@@ -1,6 +1,7 @@
-package frc.robot.autonomous;
+package frc.robot.autonomous.paths;
 
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
+import frc.robot.autonomous.FollowPath;
 import frc.robot.commandGroups.PickUpCube;
 import frc.robot.commandGroups.UpperScoring;
 import frc.robot.subsystems.arm.ArmConstants;
